+++ packages/sandbox/src/file_watching_sandbox.ts
@@ -352,6 +352,7 @@ export class FileWatchingSandbox extends EventEmitter implements Sandbox {
           'AccessDeniedException',
           'NotAuthorized',
           'ExpiredTokenException',
+          'ExpiredToken',
           'InvalidSignatureException',
         ].includes(e.name)
       ) {