+++ packages/ai-constructs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aws-amplify/ai-constructs
 
+## 0.1.4
+
+### Patch Changes
+
+- 8e964e1: fix invalid graphql in tool query generation
+
+## 0.1.3
+
+### Patch Changes
+
+- 041d041: Do not bundle AWS SDK if default implementation is used
+
 ## 0.1.2
 
 ### Patch Changes