+++ packages/backend-storage/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @aws-amplify/backend-storage
 
+## 1.1.3
+
+### Patch Changes
+
+- e648e8e: added main field to package.json so these packages are resolvable
+- Updated dependencies [8dd7286]
+  - @aws-amplify/backend-output-storage@1.1.2
+  - @aws-amplify/plugin-types@1.2.2
+
 ## 1.1.2
 
 ### Patch Changes